requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.1
PyJWT==2.9.0
pydantic-settings==2.6.1
PyYAML==6.0.2
redis==5.0.7
httpx==0.27.2

